requirements.txt
pytest>=7.0
pydantic>=2.0
tqdm>=4.60
python-dotenv>=1.0
pytimeparse>=1.1
